p4i integration Integrate 273072 from 4.3 to main: Fixes:    Complex paths with many sub-paths rendered incorrectly with blackraster. Task:     169997 RevBy:    Gunnar Details:  Prevent integer overflow when there's a lot of contours.
[git-p4: depot-paths = "//depot/qt/main/": change = 273076]

+++ src/gui/painting/qblackraster.c
@@ -1487,8 +1487,8 @@ Fail:
 /*    SUCCESS on success, FAILURE on error.                              */
 /*                                                                       */
 static Bool
-Decompose_Curve( TRaster_Instance*  raster, UShort  first,
-                 UShort  last,
+Decompose_Curve( TRaster_Instance*  raster, UInt first,
+                 UInt last,
                  int     flipped )
 {
     QT_FT_Vector   v_last;
@@ -1729,8 +1729,8 @@ Convert_Glyph( TRaster_Instance*  raster, int  flipped )
         ras.state    = Unknown_State;
         ras.gProfile = NULL;
 
-        if ( Decompose_Curve( raster, (unsigned short)start,
-                              (unsigned short)ras.outline.contours[i],
+        if ( Decompose_Curve( raster, start,
+                              ras.outline.contours[i],
                               flipped ) )
             return FAILURE;
